requirements.txt
asgiref==3.8.1
Django==5.0.4
django-allauth==0.63.3
django-environ==0.11.2
Faker==25.9.1
pillow==10.3.0
psycopg==3.1.19
psycopg2==2.9.9
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
